requirements.txt
asgiref==3.3.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cycler==0.10.0
Django==3.1.4
docopt==0.6.2
GeoAlchemy2==0.8.4
geoserver-rest==1.2.4
idna==2.10
Js2Py==0.70
kiwisolver==1.3.1
matplotlib==3.3.3
munch==2.5.0
packaging==20.7
pandas==1.1.5
Pillow==8.0.1
pipwin==0.5.0
psycopg2==2.8.6
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.2
pySmartDL==1.3.4
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.22
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.2
